--- src/pages/ChatRoom/components/MessageList.types.ts
@@ -1,5 +1,0 @@
-import {User} from 'types/chatroom.types';
-
-export interface MessageListProps {
-  usersMap?: Record<string, User>;
-}
