--- src/pages/ChatRoom/components/SendMessage.types.ts
@@ -1,1 +1,0 @@
-export interface ISendMessage {}
