--- src/pages/ChatRoom/components/UserList.types.ts
@@ -1,5 +1,0 @@
-import {User} from 'types/chatroom.types';
-
-export interface IUserList {
-  usersMap?: Record<string, User>;
-}
